+++ transformations/numeric_to_word/numeric2word.py
@@ -7,7 +7,6 @@ import numpy as np
 import phonenumbers
 from num2words import num2words
 from phonenumbers import carrier, timezone, geocoder
-from supplements import symbol_to_currency_name_dict, TimeInWords2, abbreviated_currency_symbols_to_currency_name_dict
 
 ### Recognizers
 
@@ -269,3 +268,319 @@ def recognize_transform(token):
             return token
     else:
         return token
+
+### Supplements
+
+class TimeInWords2():
+    def __init__(self):
+        self.words=["one", "two", "three", "four", "five", "six", "seven", "eight","nine", 
+       "ten", "eleven", "twelve", "thirteen", "fourteen", "quarter", "sixteen",
+       "seventeen", "eighteen", "nineteen", "twenty", "twenty one", 
+       "twenty two", "twenty three", "twenty four", "twenty five", 
+       "twenty six", "twenty seven", "twenty eight", "twenty nine", "half"]
+         
+    def caltime(self, time_string):
+        splitter = ''.join([i for i in time_string if not i.isdigit()])
+        hrs = int(time_string.split(splitter)[0])
+        mins = int(time_string.split(splitter)[1])
+        header = "" #It is..
+        msg=""
+        am_pm = "am"
+        if (hrs >12):
+            hrs=hrs-12
+            am_pm = "pm"
+        if (mins == 0):
+            hr = self.words[hrs-1]
+            msg=header + hr + " o'clock" + " " + am_pm #+ "."
+        elif (mins < 31):      
+               hr = self.words[hrs-1]
+               mn = self.words[mins-1]
+               msg = header + mn + " past " + hr + " " + am_pm #+ "."
+        else:
+            hr = self.words[hrs]
+            mn =self.words[(60 - mins-1)]
+            msg = header + mn + " to " + hr + " " + am_pm #+ "."
+        return msg
+    
+symbol_to_currency_name_dict = {"؋":"Afghanistani Afghani", 
+                                "₼":"Azerbaijani Manat",
+                                ".د.ب":"Bahraini Dinar",
+                                "৳":"Bangladeshi Taka",
+                                "Nu.":"Bhutanese Ngultrum",
+                                "$":"Dollar",
+                                "៛":"Cambodian Riel",
+                                "¥":"Chinese Yuan Renminbi", # Can also be"¥":"Japanese Yen"
+                                "ლ":"Georgian Lari",
+                                "₹":"Indian Rupee",
+                                "Rp":"Indonesian Rupiah",
+                                "﷼":"Iranian Rial",
+                                "ع.د":"Iraqi Dinar",
+                                "₪":"Israeli New Sheqel",
+                                "د.ا":"Jordanian Dinar",
+                                "лв":"Kazakhstani Tenge",
+                                "د.ك":"Kuwaiti Dinar",
+                                "лв":"Kyrgyzstani Som",
+                                "₭":"Lao Kip",
+                                "RM":"Malaysian Ringgit",
+                                "Rf":"Maldives Rufiyaa",
+                                "₮":"Mongolian Tugrik",
+                                "K":"(Burma) Myanmar Kyat",
+                                "₨":"Nepalese Rupee", 
+                                "₩":"North Korean Won",
+                                "﷼":"Omani Rial",
+                                "₨":"Pakistan Rupee",
+                                "د.ا":"Jordanian Dinar",
+                                "₱":"Philippine Peso",
+                                "﷼":"Qatari Riyal",
+                                "₽":"Russian Ruble",
+                                "﷼":"Saudi Arabian Riyal",
+                                "₩":"Korean Won",
+                                "₨":"Sri Lankan Rupee",
+                                "NT$":"New Taiwan Dollar",
+                                "ЅM":"Tajikistan Somoni",
+                                "฿":"Thai Baht",
+                                "₺":"Turkish New Lira",
+                                "T":"Turkmenistani Manat",
+                                "د.إ":"United Arab Emirates Dirham",
+                                "лв":"Uzbekistani Som",
+                                "₫":"Viet Nam Dong",
+                                "﷼":"Yemeni Rial",
+                                "دج":"Algerian Dinar",
+                                "Kz":"Angolan Kwanza",
+                                "P":"Botswana Pula",
+                                "FBu":"Burundian Franc",
+                                "FCFA":"Central African CFA",
+                                "CF":"Comorian Franc",
+                                "FC":"Congolese franc",
+                                "Fdj":"Djiboutian Franc",
+                                "ናቕፋ":"Eritrean Nakfa",
+                                "ብር":"Ethiopian Birr",
+                                "D":"Gambian Dalasi",
+                                "GH₵":"Ghanaian Cedi",
+                                "FG":"Guinean Franc",
+                                "KSh,":"Kenyan Shilling",
+                                "L":"Lesotho Loti",
+                                "ل.د":"Libyan Dinar",
+                                "Ar":"Malagasy Ariary",
+                                "MK":"Malawian Kwacha",
+                                "UM":"Mauritanian Ouguiya",
+                                "₨":"Mauritian Rupee",
+                                "DH":"Moroccan Dirham",
+                                "MT":"Mozambican Metical",
+                                "₦":"Nigerian Naira",
+                                "FRw":"Rwandan Franc",
+                                "Db":"Sao Tome Dobra",
+                                "CFA":"West African CFA",
+                                "₨":"Seychelles Rupee",
+                                "Le":"Sierra Leonean Leone",
+                                "S":"Somali Shilling",
+                                "R":"South African Rand",
+                                "SD":"Sudanese pound",
+                                "E":"Swazi Lilangeni",
+                                "TSh":"Tanzanian Shilling",
+                                "د.ت":"Tunisian Dinar",
+                                "USh":"Ugandan Shilling",
+                                "ZK":"Zambian Kwacha",
+                                "Lek":"Albanian Lek",
+                                "€":"Euro",
+                                "դր.":"Armenian Dram",
+                                "₼":"Azerbaijani Manat",
+                                "Br":"Belarusian Ruble",
+                                "KM":"Bosnia-Herzegovina Convertible Mark",
+                                "лв":"Bulgarian Lev",
+                                "kn":"Croatian Kuna",
+                                "Kč":"Czech Koruna",
+                                "kr.":"Danish Krone",
+                                "EEK":"Estonian Kroon",
+                                "ლ":"Georgian Lari",
+                                "kr":"Icelandic Krona",
+                                "Ft":"Hungarian Forint",
+                                "лв":"Kazakhstani Tenge",
+                                "Ls":"Latvian Lats",
+                                "CHF":"Swiss Franc",
+                                "Lt":"Lithuanian Litas",
+                                "ден":"Macedonian Denar",
+                                "₤":"Maltese Lira",
+                                "L":"Moldovan Leu",
+                                "kr":"Norwegian Krone",
+                                "zł":"Polish Zloty",
+                                "lei":"Romanian Leu",
+                                "₽":"Russian Ruble",
+                                "Дин.":"Serbian Dinar",
+                                "Sk":"Slovak Koruna",
+                                "kr":"Swedish Krona",
+                                "CHF":"Swiss Franc",
+                                "₺":"Turkish New Lira",
+                                "₴":"Ukrainian Hryvnia",
+                                "£":"Pound",
+                                "BZ$":"Belize Dollar",
+                                "₡":"Costa Rican Colon",
+                                "RD$":"Dominican Peso",
+                                "Q":"Guatemalan Quetzal",
+                                "G":"Haitian Gourde",
+                                "L":"Honduran Lempira",
+                                "J$":"Jamaican Dollar",
+                                "C$":"Nicaraguan Córdoba",
+                                "B/.":"Panamanian Balboa",
+                                "TT$":"Trinidad and Tobago Dollar",
+                                "$b":"Bolivian Boliviano",
+                                "Gs":"Paraguay Guarani",
+                                "S/.":"Peruvian Nuevo Sol",
+                                "$U":"Uruguayan peso",
+                                "Bs":"Venezuelan Bolivar",
+                                "K":"Papua New Guinea Kina",
+                                "WS$":"Samoan Tala",
+                                "T$":"Tongan Pa\'Anga",
+                                "VT":"Vanuatu Vatu"}
+
+abbreviated_currency_symbols_to_currency_name_dict = {'AFN': 'Afghanistani Afghani',
+                                                         'AMD': 'Armenian Dram',
+                                                         'AZN': 'Azerbaijani Manat',
+                                                         'BHD': 'Bahraini Dinar',
+                                                         'BDT': 'Bangladeshi Taka',
+                                                         'BTN': 'Bhutanese Ngultrum',
+                                                         'BND': 'Brunei Dollar',
+                                                         'KHR': 'Cambodian Riel',
+                                                         'CNY': 'Chinese Yuan Renminbi',
+                                                         'CYP': 'Cypriot Pound',
+                                                         'GEL': 'Georgian Lari',
+                                                         'INR': 'Indian Rupee',
+                                                         'IDR': 'Indonesian Rupiah',
+                                                         'IRR': 'Iranian Rial',
+                                                         'IQD': 'Iraqi Dinar',
+                                                         'ILS': 'Israeli New Sheqel',
+                                                         'JPY': 'Japanese Yen',
+                                                         'JOD': 'Jordanian Dinar',
+                                                         'KZT': 'Kazakhstani Tenge',
+                                                         'KWD': 'Kuwaiti Dinar',
+                                                         'KGS': 'Kyrgyzstani Som',
+                                                         'LAK': 'Lao Kip',
+                                                         'LBP': 'Lebanese Pound',
+                                                         'MYR': 'Malaysian Ringgit',
+                                                         'MVR': 'Maldives Rufiyaa',
+                                                         'MNT': 'Mongolian Tugrik',
+                                                         'MMK': '(Burma) Myanmar Kyat',
+                                                         'NPR': 'Nepalese Rupee',
+                                                         'KPW': 'North Korean Won',
+                                                         'OMR': 'Omani Rial',
+                                                         'PKR': 'Pakistan Rupee',
+                                                         'PHP': 'Philippine Peso',
+                                                         'QAR': 'Qatari Riyal',
+                                                         'RUB': 'Russian Ruble',
+                                                         'SAR': 'Saudi Arabian Riyal',
+                                                         'SGD': 'Singapore Dollar',
+                                                         'KRW': 'Korean Won',
+                                                         'LKR': 'Sri Lankan Rupee',
+                                                         'SYP': 'Syrian Pound',
+                                                         'TWD': 'New Taiwan Dollar',
+                                                         'TJS': 'Tajikistan Somoni',
+                                                         'THB': 'Thai Baht',
+                                                         'USD': 'United States Dollar',
+                                                         'TRY': 'Turkish New Lira',
+                                                         'TMM': 'Turkmenistani Manat',
+                                                         'AED': 'United Arab Emirates Dirham',
+                                                         'UZS': 'Uzbekistani Som',
+                                                         'VND': 'Viet Nam Dong',
+                                                         'YER': 'Yemeni Rial',
+                                                         'DZD': 'Algerian Dinar',
+                                                         'AOA': 'Angolan Kwanza',
+                                                         'XOF': 'West African CFA',
+                                                         'BWP': 'Botswana Pula',
+                                                         'BIF': 'Burundian Franc',
+                                                         'CVE': 'Cape Verde Escudo',
+                                                         'XAF': 'Central African CFA',
+                                                         'KMF': 'Comorian Franc',
+                                                         'CDF': 'Congolese franc',
+                                                         'DJF': 'Djiboutian Franc',
+                                                         'EGP': 'Egyptian Pound',
+                                                         'ERN': 'Eritrean Nakfa',
+                                                         'ETB': 'Ethiopian Birr',
+                                                         'GMD': 'Gambian Dalasi',
+                                                         'GHC': 'Ghanaian Cedi',
+                                                         'GNF': 'Guinean Franc',
+                                                         'KES': 'Kenyan Shilling',
+                                                         'LSL': 'Lesotho Loti',
+                                                         'LRD': 'Liberian Dollar',
+                                                         'LYD': 'Libyan Dinar',
+                                                         'MGA': 'Malagasy Ariary',
+                                                         'MWK': 'Malawian Kwacha',
+                                                         'MRO': 'Mauritanian Ouguiya',
+                                                         'MUR': 'Mauritian Rupee',
+                                                         'MAD': 'Moroccan Dirham',
+                                                         'MZN': 'Mozambican Metical',
+                                                         'NAD': 'Namibian Dollar',
+                                                         'NGN': 'Nigerian Naira',
+                                                         'RWF': 'Rwandan Franc',
+                                                         'STD': 'Sao Tome Dobra',
+                                                         'SCR': 'Seychelles Rupee',
+                                                         'SLL': 'Sierra Leonean Leone',
+                                                         'SOS': 'Somali Shilling',
+                                                         'ZAR': 'South African Rand',
+                                                         'SSP': 'South Sudanese pound',
+                                                         'SDG': 'Sudanese pound',
+                                                         'SZL': 'Swazi Lilangeni',
+                                                         'TZS': 'Tanzanian Shilling',
+                                                         'TND': 'Tunisian Dinar',
+                                                         'UGX': 'Ugandan Shilling',
+                                                         'ZMK': 'Zambian Kwacha',
+                                                         'ZWD': 'Zimbabwean Dollar',
+                                                         'ALL': 'Albanian Lek',
+                                                         'EUR': 'European Euro',
+                                                         'BYR': 'Belarusian Ruble',
+                                                         'BAM': 'Bosnia-Herzegovina Convertible Mark',
+                                                         'BGN': 'Bulgarian Lev',
+                                                         'HRK': 'Croatian Kuna',
+                                                         'CZK': 'Czech Koruna',
+                                                         'DKK': 'Danish Krone',
+                                                         'EEK': 'Estonian Kroon',
+                                                         'HUF': 'Hungarian Forint',
+                                                         'ISK': 'Icelandic Krona',
+                                                         'LVL': 'Latvian Lats',
+                                                         'CHF': 'Swiss Franc',
+                                                         'LTL': 'Lithuanian Litas',
+                                                         'MKD': 'Macedonian Denar',
+                                                         'MTL': 'Maltese Lira',
+                                                         'MDL': 'Moldovan Leu',
+                                                         'NOK': 'Norwegian Krone',
+                                                         'PLN': 'Polish Zloty',
+                                                         'RON': 'Romanian Leu',
+                                                         'RSD': 'Serbian Dinar',
+                                                         'SKK': 'Slovak Koruna',
+                                                         'SEK': 'Swedish Krona',
+                                                         'UAH': 'Ukrainian Hryvnia',
+                                                         'GBP': 'United Kingdom Pound Sterling',
+                                                         'XCD': 'East Caribbean Dollar',
+                                                         'BSD': 'Bahamian Dollar',
+                                                         'BBD': 'Barbados Dollar',
+                                                         'BZD': 'Belize Dollar',
+                                                         'CAD': 'Canadian Dollar',
+                                                         'CRC': 'Costa Rican Colon',
+                                                         'CUC': 'Cuban Convertible Peso',
+                                                         'DOP': 'Dominican Peso',
+                                                         'GTQ': 'Guatemalan Quetzal',
+                                                         'HTG': 'Haitian Gourde',
+                                                         'HNL': 'Honduran Lempira',
+                                                         'JMD': 'Jamaican Dollar',
+                                                         'MXN': 'Mexican Peso',
+                                                         'NIO': 'Nicaraguan Córdoba',
+                                                         'PAB': 'Panamanian Balboa',
+                                                         'TTD': 'Trinidad and Tobago Dollar',
+                                                         'ARS': 'Argentine Peso',
+                                                         'BOB': 'Bolivian Boliviano',
+                                                         'BRL': 'Brazilian Real',
+                                                         'CLP': 'Chilean Peso',
+                                                         'COP': 'Colombian Peso',
+                                                         'GYD': 'Guyanese Dollar',
+                                                         'PYG': 'Paraguay Guarani',
+                                                         'PEN': 'Peruvian Nuevo Sol',
+                                                         'SRD': 'Suriname Dollar',
+                                                         'UYU': 'Uruguayan peso',
+                                                         'VEF': 'Venezuelan Bolivar',
+                                                         'AUD': 'Australian Dollar',
+                                                         'FJD': 'Fiji Dollar',
+                                                         'NZD': 'New Zealand Dollar',
+                                                         'PGK': 'Papua New Guinea Kina',
+                                                         'WST': 'Samoan Tala',
+                                                         'SBD': 'Solomon Islands Dollar',
+                                                         'TOP': 'Tongan Pa\'Anga',
+                                                         'VUV': 'Vanuatu Vatu'}
