+++ transformations/numeric_to_word/test.json
@@ -1,49 +1,139 @@
 {
-  "type": "butter_fingers_perturbation",
+  "type": "numeric_to_word",
   "test_cases": [
     {
-      "class": "ButterFingersPerturbation",
+      "class": "NumericToWord",
       "inputs": {
-        "sentence": "Andrew finally returned the French book to Chris that I bought last week"
+        "sentence": "Please buy me 20 apples"
       },
       "outputs": {
-        "sentence": "Andgew finally returned the French book to Chrus thav I bought last week"
+        "sentence": "Please buy me twenty apples"
       }
     },
     {
-      "class": "ButterFingersPerturbation",
+      "class": "NumericToWord",
       "inputs": {
-        "sentence": "Sentences with gapping, such as Paul likes coffee and Mary tea, lack an overt predicate to indicate the relation between two or more arguments."
+        "sentence": "The deadline is in 2020/01/02"
       },
       "outputs": {
-        "sentence": "Senhences with gapping, such as Paul likes coffwe and Mary tea, lack an overt predicate to indicatx the relauion between two or more argumentd."
+        "sentence": "The deadline is in the second of january twenty twenty"
       }
     },
     {
-      "class": "ButterFingersPerturbation",
+      "class": "NumericToWord",
       "inputs": {
-        "sentence": "Alice in Wonderland is a 2010 American live-action/animated dark fantasy adventure film"
+        "sentence": "The deadline is in 2020/01"
       },
       "outputs": {
-        "sentence": "Alife in Wonderland is a 2010 American live-action/abimatev dark fantasy adventure film"
+        "sentence": "The deadline is in 2020/01"
       }
     },
     {
-      "class": "ButterFingersPerturbation",
+      "class": "NumericToWord",
       "inputs": {
-        "sentence": "Ujjal Dev Dosanjh served as 33rd Premier of British Columbia from 2000 to 2001"
+        "sentence": "The deadline is in Jan 2020"
       },
       "outputs": {
-        "sentence": "Ujjwl Dev Dosanjh served as 33rd Premier of Britush Conumbia from 2000 to 2001"
+        "sentence": "The deadline is in Jan twenty twenty"
       }
     },
     {
-      "class": "ButterFingersPerturbation",
+      "class": "NumericToWord",
       "inputs": {
-        "sentence": "Neuroplasticity is a continuous processing allowing short-term, medium-term, and long-term remodeling of the neuronosynaptic organization."
+        "sentence": "Slow down, it's still 5:00"
       },
       "outputs": {
-        "sentence": "Neugoplasticity is a continuous processing alliwing vhort-term, medium-term, and long-term remodeling of the beuronosynaptic organization."
+        "sentence": "Slow down, it's still five o'clock am"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "Quick!, it's already 23:00"
+      },
+      "outputs": {
+        "sentence": "Quick!, it's already eleven o'clock pm"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "This is 2020!"
+      },
+      "outputs": {
+        "sentence": "This is twenty twenty!"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "My phone number is +1371893178"
+      },
+      "outputs": {
+        "sentence": "My phone number is plus one three seven one eight nine three one seven eight "
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "My phone number is +6287822216501"
+      },
+      "outputs": {
+        "sentence": "My phone number is plus six two eight seven eight two two two one six five zero one "
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "My phone number is 6287822216501"
+      },
+      "outputs": {
+        "sentence": "My phone number is six two eight seven eight two two two one six five zero one "
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "The price is $300"
+      },
+      "outputs": {
+        "sentence": "The price is three hundred dollar"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "The price is 300$"
+      },
+      "outputs": {
+        "sentence": "The price is three hundred dollar"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "The price is USD300"
+      },
+      "outputs": {
+        "sentence": "The price is three hundred united states dollar"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "The price is 300USD"
+      },
+      "outputs": {
+        "sentence": "The price is three hundred united states dollar"
+      }
+    },
+    {
+      "class": "NumericToWord",
+      "inputs": {
+        "sentence": "The price is USD300!@#!"
+      },
+      "outputs": {
+        "sentence": "The price is USD300!@#!"
       }
     }
   ]
